Change Model.pickle_path from a FilePathField to a TextField
Django stores absolute paths in FilePathField, which causes problems with migrations. The absolute pickle path may change depending on the developer/environment, and this path is stored in the migrations, so any time a different developer makes migrations, a new migration is created that changes the pickle path to match the developer's local environment. We'll just use a TextField instead, where the value of the field will be interpreted relative to settings.MODEL_PICKLE_PATH.