Reimplement fill_queue to avoid concurrency issues
Perform the select and insert in a single query so they're atomic; this
prevents multiple fill_queue() calls at the same time from accessing the same
set of data and filling with duplicate data.  Also fix a few tests
that were apparently not working correctly (queue.length changes not
propagated to the database).