Make sure redis is ordered and FIFO
Do not rely on the insertion order of SELECT INSERT INTO even if the
select is ORDER BY something.  This means that what is in the DataQueue
table is ordered with respect to the rest of the data set (the MOST uncertain
objects are in the queue and LESS uncertain objects are not) the objects themselves
are not ordered.