Add Dockerfile, requirements.txt, and UsageExamples to model export
Since we are using pickles it is important to preserve the environment in
which the pickles are created.  Additionally sklearn models are picky about
the version of sklearn models are created with.  Therefore we have added a
requirements.txt file which pins the versions and a Dockerfile which can be used
to setup the environemnt.

Also U\updated the export readme to account for all the new files.